chore: move auth to stores folder & rename store files
