+++ client/src/components/navigation/sidebar.tsx
@@ -2,7 +2,7 @@ import { Trans, useLingui } from '@lingui/react/macro';
 import { type LinkProps } from '@tanstack/react-router';
 import { toast } from 'react-hot-toast';
 
-import { logout, useAuthStore } from '~/services/auth';
+import { logout, useAuthStore } from '~/stores/auth-store';
 import { styled } from '~/styled-system/jsx';
 import { stack } from '~/styled-system/patterns';
 import { Button } from '~/uikit/button';
