+++ client/src/stores/workspace-store.ts
